+++ src/helper/getCurrentFilePath.js
@@ -1,0 +1,3 @@
+import url from "url";
+
+export default (meta) => url.fileURLToPath(meta.url);
